// commerical calculator/types.ts
export interface QuotationDraft {
  clientname: string;
  clientcompany: string;
  contactno: string;
  email: string;
  addressline1: string;
  addressline2: string;
  addressline3: string;
  systemsize: string;
  panel: string;
  invertersize: string;
  inverterbrand: string;
  systemprice: number;
  date: string;
}